As part of the Nuke integration, we can no longer guarantee that included/required files are in the current directory.  So, we now prepend $GALLERY_BASEDIR to all included/required files.
There's a bug in PHP 4.0.1pl2 that require_once() barfs if you
give it a variable instead of a constant, so we now use require()
instead.  Once upon a time when we had init.php set up as an
auto_prepend, require_once was more valuable for us.  For now,
it's not really necessary.

Don't emit the <body> and <head> tags if we're embedded inside
another application.

do_command.php links are now funneled through the doCommand() function
which rewrites them appropriately depending on whether Gallery is
running standalone or embedded.

All gallery/album urls are funneled through makeGalleryUrl()


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery@771 57fcd75e-5312-0410-8df3-f5eb6fbb1595