Some small changes:
* Only display "* comments available" text at the bottom
of an album if there was actually a comment for a photo
on that page.

* Only display "Viewed n time(s)" if a photo has actually been
viewed before.

* Fixed bug in [reset counter] functionality in do_command.php
where it wouldn't work for a non-admin user because $gallery->admin
wasnt' defined yet.

* Added default settings for photo commenting and click counting
in the setup wizard.  Incremented version.php

* Search will search comments if they exist for a given photo.
A given comment will only be displayed in the search engine
if it matches.

-jpk


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery@849 57fcd75e-5312-0410-8df3-f5eb6fbb1595