1.3.5-cvs-b107
Updated all refs to 1.3.5 to 1.4, except version number.

Added filesize restriction for resize, and bump album version to
store this value for each album.  If this value is set, the compress
function will be called repeatedly to get as close as possible to
this size.
		- changes to move_photo.php resizing

Stop language initialition in embedded galleries.

Fixed bug that print statement in move_photo.php

Fixed bug in displaying number hidden albums in view album (thanks
to Eelco Brolman)

Change "configurations wizard" to "configuration wizard".


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery@3547 57fcd75e-5312-0410-8df3-f5eb6fbb1595